AR-618: moved files into shared folder
