+++ src/app.jsx
@@ -8,11 +8,11 @@ import Col from "react-bootstrap/Col";
 import TagList from "./components/tag-list/tag-list";
 import TopBar from "./components/navigation/topbar/topbar";
 import SideBar from "./components/navigation/sidebar/sidebar";
-import ScreenList from "./components/screen-list/screen-list";
+import ScreenList from "./components/screen/screen-list";
 import CategoryList from "./components/category-list/category-list";
 import SlidesList from "./components/slides-list/slides-list";
 import EditTag from "./components/edit-tag/edit-tag";
-import EditScreen from "./components/edit-screen/edit-screen";
+import EditScreen from "./components/screen/edit-screen";
 import EditCategories from "./components/edit-categories/edit-category";
 import GroupsList from "./components/groups-list/groups-list";
 import EditGroup from "./components/edit-group/edit-group";
