+++ src/components/screen/grid-generation-and-select.jsx
